Allow to sort files according to Exif-date
I also like the idea about sorting files according to Exif-date. This
behaviour seemed to be implemented in the pan-view but not in the main
browser view. I created a little patch to "correct" this issue.  ;)

This satisfies my own needs, although the exif-date reading could be
probably somehow optimised(?).. now the GUI becomes unresponsive for
few seconds if there are a lot of pictures in the directory...